also move ClassLoadingUtils
