+++ flink-rpc/flink-rpc-akka/src/main/java/org/apache/flink/runtime/concurrent/akka/ActorSystemScheduledExecutorAdapter.java
@@ -18,6 +18,7 @@
 
 package org.apache.flink.runtime.concurrent.akka;
 
+import org.apache.flink.runtime.concurrent.ClassLoadingUtils;
 import org.apache.flink.util.Preconditions;
 import org.apache.flink.util.concurrent.ScheduledExecutor;
